requirements.txt
boto3==1.42.30
deepdish==0.3.7
imageio==2.37.0
matplotlib==3.10.8
memory_profiler==0.61.0
mmcv==2.2.0
mpi4py==4.1.1
nose==1.3.7
numba==0.63.1
numpy==1.23.5
packaging==25.0
pandas==1.5.3
Pillow==12.1.0
pycocotools==2.0.11
pydaily==0.4.4
scipy==1.7.3
setuptools==80.9.0
Shapely==2.1.2
skimage==0.0
slide==0.0.2
smmap==5.0.2
tensorboardx==2.6.4
tifffile==2024.8.30
torch==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.67.1
